src/mining/requirements.txt
gitpython==3.1.30
PyDriller==2.5
requests==2.25.1
nltk
pyyaml
networkx
